test util cleanup
