+++ Tests/Fluent/Utilities/Atom.swift
@@ -3,25 +3,24 @@ import Fluent
 final class Atom: Entity {
     var id: Node?
     var name: String
-    var protons: Int
+    var groupId: Node?
 
-    init(name: String, protons: Int) {
+    init(name: String) {
         id = nil
         self.name = name
-        self.protons = protons
     }
 
     init(with node: Node, in context: Context) throws {
         id = try node.extract("id")
         name = try node.extract("name")
-        protons = try node.extract("protons")
+        groupId = try node.extract("group_id")
     }
 
     func makeNode() throws -> Node {
         return try Node([
             "id": id,
             "name": name,
-            "protons": protons
+            "group": try group().get()
         ])
     }
 
@@ -29,11 +28,23 @@ final class Atom: Entity {
         return try siblings()
     }
 
+    func group() throws -> Parent<Group> {
+        return try parent(groupId)
+    }
+
+    func protons() throws -> Children<Proton> {
+        return children()
+    }
+
+    func nucleus() throws -> Nucleus? {
+        return try children().first()
+    }
+
     static func prepare(_ database: Fluent.Database) throws {
         try database.create(entity) { builder in
             builder.id()
             builder.string("name")
-            builder.int("protons")
+            builder.int("group_id")
         }
     }
     static func revert(_ database: Fluent.Database) throws {
